protoype working return pv.DataSets and pv.PolyData
