+++ src/ansys/dpf/core/plotter.py
@@ -12,7 +12,6 @@ import tempfile
 import os
 import sys
 import numpy as np
-import inspect
 import warnings
 from typing import TYPE_CHECKING, List, Union
 
@@ -20,37 +19,13 @@ from ansys import dpf
 from ansys.dpf import core
 from ansys.dpf.core.common import locations, DefinitionLabels
 from ansys.dpf.core.common import shell_layers as eshell_layers
+from ansys.dpf.core.helpers import _sort_supported_kwargs
 from ansys.dpf.core import errors as dpf_errors
+from ansys.dpf.core.helpers import compute_streamlines
 from ansys.dpf.core.nodes import Node, Nodes
 
 if TYPE_CHECKING:  # pragma: no cover
     from ansys.dpf.core.meshed_region import MeshedRegion
-
-
-def _sort_supported_kwargs(bound_method, **kwargs):
-    """Filters the kwargs for a given method."""
-    # Ignore warnings unless specified
-    if not sys.warnoptions:
-        import warnings
-
-        warnings.simplefilter("ignore")
-    # Get supported arguments
-    supported_args = inspect.getfullargspec(bound_method).args
-    kwargs_in = {}
-    kwargs_not_avail = {}
-    # Filter the given arguments
-    for key, item in kwargs.items():
-        if key in supported_args:
-            kwargs_in[key] = item
-        else:
-            kwargs_not_avail[key] = item
-    # Prompt a warning for arguments filtered out
-    if len(kwargs_not_avail) > 0:
-        txt = f"The following arguments are not supported by {bound_method}: "
-        txt += str(kwargs_not_avail)
-        warnings.warn(txt)
-    # Return the accepted arguments
-    return kwargs_in
 
 
 class _InternalPlotterFactory:
@@ -334,51 +309,36 @@ class _PyVistaPlotter:
                 point_size=label_point_size,
             )
 
-    def add_streamlines(self, meshed_region, field, computed_streamlines=None, radius=1.0, **kwargs):
-        # Check velocity field location
-        if field.location is not dpf.core.locations.nodal:
-            warnings.warn(
-                "Velocity field must have a nodal location. Result must be carefully checked."
-            )
-
-        # handles input data
-        f_name = field.name
-        stream_name = "streamlines " + f_name + " (" + str(field.unit) + ")"
-        grid = meshed_region.grid
-        mesh_nodes = meshed_region.nodes
-
-        ind, mask = mesh_nodes.map_scoping(field.scoping)
-        overall_data = np.full((len(mesh_nodes), 3), np.nan)  # velocity has 3 components
-        overall_data[ind] = field.data[mask]
-
-        grid.set_active_scalars(None)
-        grid[f"{stream_name}"] = overall_data
-
-        # check src request
-        return_source = kwargs.pop("return_source", None)
+    def add_streamlines(self, meshed_region=None, field=None, computed_streamlines=None, computed_source=None, radius=1.0, **kwargs):
         permissive = kwargs.pop("permissive", None)
+        return_source = kwargs.get("return_source", None)
 
-        # filter kwargs
-        kwargs_base = _sort_supported_kwargs(bound_method=grid.streamlines, **kwargs)
-        kwargs_from_source = _sort_supported_kwargs(
-            bound_method=grid.streamlines_from_source, **kwargs
-        )
-        kwargs_from_source.update(kwargs_base)  # merge both dicts in kwargs_from_source
+        if computed_streamlines is None:
+            if meshed_region is None or field is None:
+                raise AttributeError(
+                    "if computed_streamlines argument is None,"
+                    "meshed_region and field arguments must be provided"
+                )
 
         # create streamlines
-        if computed_streamlines is not None:
-            pass
         if return_source:
-            streamlines, src = grid.streamlines(
-                vectors=f"{stream_name}",
-                return_source=True,
-                **kwargs_from_source,
-            )
+            if computed_streamlines is not None:
+                streamlines, src = computed_streamlines
+            else:
+                streamlines, src = compute_streamlines(
+                    meshed_region=meshed_region,
+                    field=field,
+                    **kwargs
+                )
         else:
-            streamlines = grid.streamlines(
-                vectors=f"{stream_name}",
-                **kwargs_from_source,
-            )
+            if computed_streamlines is not None:
+                streamlines = computed_streamlines
+            else:
+                streamlines = compute_streamlines(
+                    meshed_region=meshed_region,
+                    field=field,
+                    **kwargs
+                )
 
         # set streamline on plotter
         sargs = dict(vertical=False)
@@ -556,8 +516,8 @@ class DpfPlotter:
 
     def add_streamlines(
         self,
-        meshed_region,
-        field,
+        meshed_region=None,
+        field=None,
         computed_streamlines=None,
         radius=0.1,
         **kwargs,
@@ -573,14 +533,17 @@ class DpfPlotter:
         ----------
         meshed_region : MeshedRegion
             MeshedRegion the streamline will be computed on.
+            Not needed if computed_streamlines is provided.
         field : Field
             Field containing raw vector data the streamline is
             computed from. The data location must be nodal, velocity
             values must be defined at nodes.
+            Not needed if computed_streamlines is provided.
         computed_streamlines : FieldsContainer
             FieldsContianer containing computed streamlines data,
             computde using `dpf.helpers.compute_streamlines`
-            function.
+            function. Must be provided if field and meshed_region
+            are not provided.
         **kwargs : optional
             Additional keyword arguments for the plotter. More information
             is available at :func:`pyvista.DataSetFilters.streamlines`.
